doc updates and organization
