+++ lib/src/signals/logic_net.dart
@@ -1,41 +1,16 @@
-//TODO:header
+// Copyright (C) 2024 Intel Corporation
+// SPDX-License-Identifier: BSD-3-Clause
+//
+// logic_net.dart
+// Definition for `LogicNet`.
+//
+// 2024 May 30
+// Author: Max Korbel <max.korbel@intel.com>
 
 part of 'signals.dart';
 
-class _WireNet extends _Wire {
-  final Set<Logic> _drivers = {};
-
-  _WireNet({required super.width});
-
-  void _evaluateNewValue({required String signalName}) {
-    var newValue = LogicValue.filled(width, LogicValue.z);
-    for (final driver in _drivers) {
-      newValue = newValue.triState(driver.value);
-    }
-    put(newValue, signalName: signalName);
-  }
-
-  @override
-  void _adopt(_Wire other) {
-    assert(other is _WireNet, 'Only should be adopting other `_WireNet`s');
-    other as _WireNet;
-
-    super._adopt(other);
-
-    other._drivers
-      ..forEach(_addDriver)
-      ..clear();
-  }
-
-  void _addDriver(Logic driver) {
-    if (_drivers.add(driver)) {
-      driver.glitch.listen((args) {
-        _evaluateNewValue(signalName: driver.name);
-      });
-    }
-  }
-}
-
+/// Represents a [Logic] which can be have multiple drivers, supporting
+/// bidirectional signals.
 class LogicNet extends Logic {
   @override
   Logic? get srcConnection => null;
@@ -43,9 +18,21 @@ class LogicNet extends Logic {
   @override
   bool get isNet => true;
 
+  /// Constructs a new [LogicNet] supporting multiple drivers named [name] with
+  /// [width] bits.
+  ///
+  /// The default value for [width] is 1.  The [name] should be sanitary
+  /// (variable rules for languages such as SystemVerilog).
+  ///
+  /// The [naming] and [name], if unspecified, are chosen based on the rules in
+  /// [Naming.chooseNaming] and [Naming.chooseName], respectively.
   LogicNet({super.name, super.width, super.naming})
       : super._(wire: _WireNet(width: width));
 
+  /// Constructs a new [LogicNet] with some additional validation for ports of
+  /// [Module]s.
+  ///
+  /// Useful for [Interface] definitions.
   factory LogicNet.port(String name, [int width = 1]) {
     if (!Sanitizer.isSanitary(name)) {
       throw InvalidPortNameException(name);
