+++ lib/src/signals/signals.dart
@@ -19,6 +19,7 @@ export 'port.dart';
 part 'const.dart';
 part 'logic.dart';
 part 'wire.dart';
+part 'wire_net.dart';
 part 'logic_structure.dart';
 part 'logic_array.dart';
 part 'logic_net.dart';
